Fixed incorrect sizeof in LinkReceiver::ReadRegion.
The type of BRegion::fCount is long, and ServerLink sends/receives it
with sizeof(long), but LinkReceiver was using sizeof(int32). Due to
long being 64 bit this was resulting in a mismatch. This fixes the
drawing problems on x86_64.

+++ src/kits/app/LinkReceiver.cpp
@@ -443,7 +443,7 @@ err:
 status_t
 LinkReceiver::ReadRegion(BRegion* region)
 {
-	status_t status = Read(&region->fCount, sizeof(int32));
+	status_t status = Read(&region->fCount, sizeof(long));
 	if (status >= B_OK)
 		status = Read(&region->fBounds, sizeof(clipping_rect));
 	if (status >= B_OK) {
